refactor: 소모임 가입 패키지 위치 조정
- entity, service 디렉토리 추가 후 이동
- #26
