+++ src/main/java/com/example/wegather/smallGroupJoin/domin/repository/SmallGroupMemberRepository.java
@@ -1,6 +1,6 @@
 package com.example.wegather.smallGroupJoin.domin.repository;
 
-import com.example.wegather.smallGroupJoin.domin.SmallGroupMember;
+import com.example.wegather.smallGroupJoin.domin.entity.SmallGroupMember;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
